requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.9
django-graphql-jwt==0.3.4
django-phonenumber-field==6.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
phonenumberslite==8.12.37
promise==2.3
PyJWT==2.3.0
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
